Improve Spawner Speed by Batching Requests
